split models and db from init
Move models into their own file
Move db functions into their own file
Update tests
